fix: drop lodash
As the years passed we used lodash less and less. Before this commit, lodash was
used only for ``_.extend``, which can be replicated by ``Object.assign`` in the
usages we made of it (or even less than that).

This is not considered a breaking change as a) wed does not provide access to
lodash through its API, b) depending on lodash being distributed with wed
amounts to depending on an undocummented aspect of wed. Unsupported use.

Dropping lodash resulted in a ~10% drop in ``wed.min.js`` size and a ~20% drop
in ``wed.js`` size.