refactor: use rangy-glue directly instead of using mappings
The mappings required to cause ``import "rangy"`` to load rangy-glue instead
were a pain to maintain. Since we use rangy only for a single purpose (text
search) there's no great benefit to maintaining the mappings and be able to
import rangy with the right modules loaded automagically. So we dropped the
mappings and explicitly load rangy-glue.