fix: remove the connectivity check from runtime's fetch
That connectivity check was problematic on multiple fronts:

1. It would take priority over whatever check would have been performed by the
caller of ``fetch``. (The caller's ``catch`` would necessarily have been after
the ``catch`` local to the ``fetch``.)

2. It was generic to the point of being useless. There's not one-size-fits all
for recovering from a connectivity problem. The Ajax saver already has a way to
retry the save. Other users need to come up with their own intelligent strategy.

3. The existing code depended on Bootstrap and jQuery. It is preferable to have
the runtime be independent of these. This could have been fixed but the issues
above made it so that it was better to just remove it.