fix: handle CData as text
Recent changes were made with the hope that CData could be handled as CData when
editing XML documents with wed, but that was a naive hope. This commit makes it
so that CData is conveted to Text nodes when loading a document. It also
documents why we don't support CData.